feat: support for dynamic OpenAPI doc generation
