+++ noscrum/noscrum_api/search.py
@@ -1,14 +1,16 @@
 """
 Handle search logic
 """
-from flask import Blueprint, render_template, request, abort
-from flask_user import current_user, login_required
-from noscrum.db import get_db
+from flask_openapi3 import APIBlueprint as Blueprint
+from flask import request, abort
+from flask_login import current_user, login_required
+from noscrum.noscrum_backend.db import get_db
+from noscrum.noscrum_api.template_friendly import friendly_render as render_template
 
 bp = Blueprint("search", __name__, url_prefix="/search")
 
 
-def search_db(search_term):
+def search_db(current_user, search_term: str):
     app_db = get_db()
     alltext = """select 'task' label,task value, task.id, user_id from task
     union all select 'story',story value,story.id, user_id from story
@@ -20,12 +22,12 @@ def search_db(search_term):
     )
 
 
-@bp.route("/", methods=["GET"])
+@bp.get("/")
 @login_required
 def query():
     search_term = request.args.get("s")
     if not query:
         abort(401)
     else:
-        results = search_db(search_term)
+        results = search_db(current_user, search_term)
     return render_template("search/results.html", results=results)
