+++ noscrum/noscrum_api/work.py
@@ -1,148 +1,36 @@
 """
 Data handler for work view and controller
 """
-import json
 from datetime import date, datetime
 
-from flask import Blueprint, redirect, render_template, request, url_for, abort, flash
-from flask_user import current_user
-
-from noscrum.db import get_db, Work, ScheduleTask
-from noscrum.task import (
+from flask_openapi3 import APIBlueprint as Blueprint
+from flask import redirect, request, url_for, abort, flash
+from flask_login import current_user, login_required
+from pydantic import BaseModel, Field
+from noscrum.noscrum_api.template_friendly import friendly_render as render_template
+from noscrum.noscrum_backend.task import (
     get_task,
-    update_task,
     get_tasks_for_story,
     get_tasks_for_epic,
     get_tasks,
 )
-from noscrum.story import get_story
-from noscrum.epic import get_epic
-from noscrum.sprint import get_sprint_by_date
+from noscrum.noscrum_backend.story import get_story
+from noscrum.noscrum_backend.epic import get_epic
+from noscrum.noscrum_backend.work import *
+from noscrum.noscrum_api.task import TaskPath
+from noscrum.noscrum_api.story import StoryPath
+from noscrum.noscrum_api.epic import EpicPath
+from noscrum.noscrum_api.template_friendly import NoscrumBaseQuery
 
 bp = Blueprint("work", __name__, url_prefix="/work")
 
 
-# pylint: disable-next=too-many-arguments
-def create_work(work_date, hours_worked, status, task_id, new_actual, update_status):
-    """
-    Create new work for task on the given date
-    @param work_date when some action was done
-    @param hours_worked number of hours worked
-    @param status a new task status given work
-    @param task_id task which work is executed
-    @param new_actual task actual hours worked
-    @param update_status boolean update status
-    """
-    app_db = get_db()
-    work_schedule = (
-        ScheduleTask.query.filter(ScheduleTask.sprint_day == work_date)
-        .filter(ScheduleTask.task_id == task_id)
-        .filter(ScheduleTask.user_id == current_user.id)
-        .first()
-    )
-    if work_schedule is None:
-        raise Exception("Cannot log work on unscheduled task")
-    new_work = Work(
-        work_date=work_date,
-        hours_worked=hours_worked,
-        status=status,
-        task_id=task_id,
-        user_id=current_user.id,
-    )
-    app_db.session.add(new_work)
-    new_status = status if update_status else None
-    update_task(task_id, None, None, None, new_status, new_actual, None, None, None)
-    app_db.session.commit()
-
-
-def get_work(work_id):
-    """
-    Get work record from identification number
-    @param work_id work record identity number
-    """
-    return (
-        Work.query.filter(Work.id == work_id)
-        .filter(Work.user_id == current_user.id)
-        .first()
-    )
-
-
-def get_work_for_task(task_id):
-    """
-    Get the work records given the task record
-    @param task_id task record work is queried
-    """
-    return (
-        Work.query.filter(Work.task_id == task_id)
-        .filter(Work.user_id == current_user.id)
-        .all()
-    )
-
-
-def get_work_for_story(story_id):
-    """
-    Get the work for a particular story record
-    @param story_id a Story record locator val
-    """
-    return (
-        Work.query.filter(Work.story.id == story_id)
-        .filter(Work.user_id == current_user.id)
-        .all()
-    )
-
-
-def get_work_for_epic(epic_id):
-    """
-    Get all work records under the epic record
-    @param epic_id epic for which work queried
-    """
-    app_db = get_db()
-    return Work(
-        *app_db.session.execute(
-            "SELECT work.id, task_id, work_date, hours_worked, status "
-            + "FROM work JOIN task on work.task_id = task.id "
-            + "JOIN story ON task.story_id = story.id "
-            + " WHERE story.epic_id = ? order by work_date",
-            (epic_id,),
-        ).fetchall()
-    )
-
-
-def get_work_by_dates(start_date, end_date):
-    """
-    Get work executed between two dates values
-    @param start_date date request lower limit
-    @param end_date date requested upper limit
-    """
-    return Work.query.filter(Work.work_date >= start_date)\
-                .filter(Work.work_date <= end_date)\
-                .filter(Work.user_id == current_user.id)\
-            .all()
-
-def delete_work(work_id):
-    """
-    Delete work record given a certain identiy
-    @param work_id a work record to be deleted
-    """
-    app_db = get_db()
-    work = get_work(work_id)
-    Work.query.filter(Work.id == work_id).filter(
-        Work.user_id == current_user.id
-    ).delete()
-    app_db.session.commit()
-    return work.id
-
-
-@bp.route("/create/<int:task_id>", methods=("POST", "GET"))
-def create(task_id):
-    """
-    Handle requests to create work on the task
-    GET: Return form to create new work record
-    POST: Create new work record for some task
-    @param task_id task which work is executed
-    """
-    is_json = request.args.get("is_json", False)
-    task = get_task(task_id)
+@bp.get("/create/<int:task_id>")
+@login_required
+def get_create(path: TaskPath, query: NoscrumBaseQuery):
+    is_json = query.is_json
+    task_id = path.task_id
+    task = get_task(current_user, task_id)
     if task is None:
         error = f"Task {task_id} Not Found"
         if is_json:
@@ -150,8 +38,28 @@ def create(task_id):
         else:
             flash(error, "error")
             return redirect(url_for("task.list_all"))
-    if request.method == "GET":
-        return render_template("work/create.html", task=task)
+    return render_template("work/create.html", task=task)
+
+
+@bp.post("/create/<int:task_id>")
+@login_required
+def create(path: TaskPath, query: NoscrumBaseQuery):
+    """
+    Handle requests to create work on the task
+    GET: Return form to create new work record
+    POST: Create new work record for some task
+    @param task_id task which work is executed
+    """
+    is_json = query.is_json
+    task_id = path.task_id
+    task = get_task(current_user, task_id)
+    if task is None:
+        error = f"Task {task_id} Not Found"
+        if is_json:
+            abort(404, error)
+        else:
+            flash(error, "error")
+            return redirect(url_for("task.list_all"))
     # Handle POST Request
     work_date = request.form.get("work_date", date.today())
     if isinstance(work_date, str):
@@ -167,24 +75,32 @@ def create(task_id):
     status = request.form.get("status", task.status)
     update_status = request.form.get("update_status", False)
     update_status = True if update_status or update_status == "on" else False
-    true_actual = sum([x.hours_worked for x in get_work_for_task(task_id)])
+    true_actual = sum(
+        [x.hours_worked for x in get_work_for_task(current_user, task_id)]
+    )
     new_actual = hours_worked if true_actual is None else true_actual + hours_worked
-    create_work(work_date, hours_worked, status, task_id, new_actual, update_status)
+    create_work(
+        current_user,
+        work_date,
+        hours_worked,
+        status,
+        task_id,
+        new_actual,
+        update_status,
+    )
     if is_json:
         return {"Success": True, "task_id": task_id}
     return redirect(url_for("work.list_for_task", task_id=task_id))
 
+class WorkPath(BaseModel):
+    work_id: int
 
-@bp.route("/<int:work_id>", methods=("GET", "DELETE"))
-def read_delete(work_id):
-    """
-    Handle read or deletes on some work record
-    GET: Information regarding specific record
-    DELETE: Delete work record with identifier
-    @param work_id work record identifier code
-    """
-    is_json = request.args.get("is_json", False)
-    work_item = get_work(work_id)
+@bp.get("/<int:work_id>")
+@login_required
+def show(path: WorkPath, query: NoscrumBaseQuery):
+    is_json = query.is_json
+    work_id = path.work_id
+    work_item = get_work(current_user, work_id)
     if work_item is None:
         error = f"Work Item {work_id} Not Found"
         if is_json:
@@ -192,28 +108,47 @@ def read_delete(work_id):
         else:
             flash(error, "error")
             return redirect(url_for("sprint.active"))
-    if request.method == "GET":
-        if is_json:
-            return json.dumps(
-                {"Success": True, "work_id": work_id, "work_item": work_item.to_dict()}
-            )
-        return render_template("work/read_del", work_item=work_item)
-    elif request.method == "DELETE":
-        deleted_work_id = delete_work(work_id)
-        if is_json:
-            return json.dumps({"Success": True, "work_id": deleted_work_id})
-        return redirect(url_for("work.list_for_task", task_id=work_item.task_id))
+    if is_json:
+        return {"Success": True, "work_id": work_id, "work_item": work_item.to_dict()}
+    return render_template("work/read_del", work_item=work_item)
 
 
-@bp.route("/list/task/<int:task_id>", methods=("GET",))
-def list_for_task(task_id):
+@bp.route("/<int:work_id>", methods=("GET", "DELETE"))
+@login_required
+def delete(path: WorkPath, query: NoscrumBaseQuery):
+    """
+    Handle read or deletes on some work record
+    GET: Information regarding specific record
+    DELETE: Delete work record with identifier
+    @param work_id work record identifier code
+    """
+    is_json = query.is_json
+    work_id = path.work_id
+    work_item = get_work(current_user, work_id)
+    if work_item is None:
+        error = f"Work Item {work_id} Not Found"
+        if is_json:
+            abort(404, error)
+        else:
+            flash(error, "error")
+            return redirect(url_for("sprint.active"))
+    deleted_work_id = delete_work(current_user, work_id)
+    if is_json:
+        return {"Success": True, "work_id": deleted_work_id}
+    return redirect(url_for("work.list_for_task", task_id=work_item.task_id))
+
+
+@bp.get("/list/task/<int:task_id>")
+@login_required
+def list_for_task(path: TaskPath, query: NoscrumBaseQuery):
     """
     Return all work records for the given task
     GET: route provides the response described
     @param task_id task record was executed on
     """
-    is_json = request.args.get("is_json", False)
-    tasks = get_task(task_id)
+    is_json = query.is_json
+    task_id = path.task_id
+    tasks = get_task(current_user, task_id)
     if tasks is None:
         error = f"Task Item {task_id} not found"
         if is_json:
@@ -221,7 +156,7 @@ def list_for_task(task_id):
         else:
             flash(error, "error")
             return redirect(url_for("sprint.active"))
-    work_items = get_work_for_task(task_id)
+    work_items = get_work_for_task(current_user, task_id)
     if work_items is None:
         error = f"No Work Items found for Task {task_id}"
         if is_json:
@@ -231,21 +166,23 @@ def list_for_task(task_id):
             return redirect(url_for("sprint.active"))
     if is_json:
         dict_work_items = [x.to_dict() for x in work_items]
-        return json.dumps({"Success": True, "work_items": dict_work_items})
+        return {"Success": True, "work_items": dict_work_items}
     return render_template(
         "work/list.html", key="Task", tasks=tasks, work_items=work_items
     )
 
 
-@bp.route("/list/story/<int:story_id>", methods=("GET",))
-def list_for_story(story_id):
+@bp.get("/list/story/<int:story_id>")
+@login_required
+def list_for_story(path: StoryPath, query: NoscrumBaseQuery):
     """
     List all work completed on tasks for story
     GET: route provides the response described
     @param story_id story where work described
     """
-    is_json = request.args.get("is_json", False)
-    story = get_story(story_id)
+    is_json = query.is_json
+    story_id = path.story_id
+    story = get_story(current_user, story_id)
     if story is None:
         error = "Story Item not found"
         if is_json:
@@ -253,7 +190,7 @@ def list_for_story(story_id):
         else:
             flash(error, "error")
             return redirect(url_for("sprint.active"))
-    tasks = get_tasks_for_story(story_id)
+    tasks = get_tasks_for_story(current_user, story_id)
     if tasks is None:
         error = f"No Tasks found for Story {story.id}"
         if is_json:
@@ -261,7 +198,7 @@ def list_for_story(story_id):
         else:
             flash(error, "error")
             return redirect(url_for("sprint.active"))
-    work_items = get_work_for_story(story_id)
+    work_items = get_work_for_story(current_user, story_id)
     if work_items is None:
         error = f"No Work Items found for Story {story.id}"
         if is_json:
@@ -271,7 +208,7 @@ def list_for_story(story_id):
             return redirect(url_for("sprint.active"))
     if is_json:
         dict_work_items = [x.to_dict() for x in work_items]
-        return json.dumps({"Success": True, "work_items": dict_work_items})
+        return {"Success": True, "work_items": dict_work_items}
     return render_template(
         "work/list.html",
         key="Story " + story["story"],
@@ -280,15 +217,17 @@ def list_for_story(story_id):
     )
 
 
-@bp.route("/list/epic/<int:epic_id>", methods=("GET",))
-def list_for_epic(epic_id):
+@bp.get("/list/epic/<int:epic_id>")
+@login_required
+def list_for_epic(path: EpicPath, query: NoscrumBaseQuery):
     """
     List work completed on tasks an epic holds
     GET: route provides the response described
     @param epic_id record identity for an epic
     """
-    is_json = request.args.get("is_json", False)
-    epic = get_epic(epic_id)
+    epic_id = path.epic_id
+    is_json = query.is_json
+    epic = get_epic(current_user, epic_id)
     if epic is None:
         error = "Epic Item not found"
         if is_json:
@@ -296,7 +235,7 @@ def list_for_epic(epic_id):
         else:
             flash(error, "error")
             return redirect(url_for("sprint.active"))
-    tasks = get_tasks_for_epic(epic_id)
+    tasks = get_tasks_for_epic(current_user, epic_id)
     if tasks is None:
         error = "No Tasks found for Epic {epic.id}"
         if is_json:
@@ -304,7 +243,7 @@ def list_for_epic(epic_id):
         else:
             flash(error, "error")
             return redirect(url_for("sprint.active"))
-    work_items = get_work_for_epic(epic_id)
+    work_items = get_work_for_epic(current_user, epic_id)
     if work_items is None:
         error = f"No Work Items found for Epic {epic.id}"
         if is_json:
@@ -314,13 +253,14 @@ def list_for_epic(epic_id):
             return redirect(url_for("sprint.active"))
     if is_json:
         dict_work_items = [x.to_dict() for x in work_items]
-        return json.dumps({"Success": True, "work_items": dict_work_items})
+        return {"Success": True, "work_items": dict_work_items}
     return render_template(
         "work/list.html", key="Epic " + epic["epic"], tasks=tasks, work_items=work_items
     )
 
 
-@bp.route("/list/dates", methods=("GET",))
+@bp.get("/list/dates")
+@login_required
 def list_for_dates():
     """
     List all work completed within given dates
@@ -329,7 +269,7 @@ def list_for_dates():
     is_json = request.args.get("is_json", False)
     start_date = request.args.get("start_date", date(2020, 1, 1))
     end_date = request.args.get("end_date", date.today())
-    work_items = get_work_by_dates(start_date, end_date)
+    work_items = get_work_by_dates(current_user, start_date, end_date)
     if work_items is None:
         error = "No work items found between dates_provided"
         if is_json:
@@ -337,7 +277,9 @@ def list_for_dates():
         else:
             flash(error, "error")
             return redirect(url_for("sprint.active"))
-    all_tasks = get_tasks()
+    all_tasks = get_tasks(
+        current_user,
+    )
     task_ids = [x.task_id for x in work_items]
     tasks = []
     for task in all_tasks:
@@ -345,7 +287,7 @@ def list_for_dates():
             tasks.append(task)
     if is_json:
         dict_work_items = [x.to_dict() for x in work_items]
-        return json.dumps({"Success": True, "work_items": dict_work_items})
+        return {"Success": True, "work_items": dict_work_items}
     return render_template(
         "work/list.html",
         key=f"Dates from {start_date} to {end_date}",
@@ -353,7 +295,9 @@ def list_for_dates():
         work_items=work_items,
     )
 
-@bp.route("/reporting", methods=("GET",))
+
+@bp.get("/reporting")
+@login_required
 def display_report_page():
     """
     Render report page.
