+++ noscrum/noscrum_backend/extract.py
@@ -1,0 +1,20 @@
+"""
+Extract and Import data for current user.
+"""
+from noscrum.noscrum_backend.epic import get_epics
+from noscrum.noscrum_backend.story import get_stories
+from noscrum.noscrum_backend.task import get_tasks
+from noscrum.noscrum_backend.sprint import get_sprints, get_schedules
+from noscrum.noscrum_backend.tag import get_tags
+from noscrum.noscrum_backend.user import get_preferences
+
+
+def get_user_data(current_user):
+    user_export_data = {}
+    user_export_data["epic"] = get_epics(current_user)
+    user_export_data["story"] = get_stories(current_user)
+    user_export_data["task"] = get_tasks(current_user)
+    user_export_data["sprint"] = get_sprints(current_user)
+    user_export_data["schedule"] = get_schedules(current_user)
+    user_export_data["tags"] = get_tags(current_user)
+    user_export_data["preferences"] = get_preferences(current_user)
