+++ noscrum/templates/base.html
@@ -19,7 +19,7 @@
         <li><a href="{{ url_for('sprint.list_all') }}">All Sprints</a></li>
             <li><a href="{{ url_for('task.list_all') }}?archive=true">Archive</a></li>
         {% else %}
-        <li><a href="{{ url_for('user.register') }}">Register for Free</a></li> 
+        <li><a href="{{ url_for('user.get_create') }}">Register for Free</a></li> 
         {% endif %}
         <li><input type="search" name="s" placeholder="Search"></li>
         <li><button  class="button" type="submit">Search</button></li>
@@ -27,11 +27,11 @@
     </form>
         <div class="float-right">
             {% if current_user.is_authenticated %}
-                <a href="{{ url_for('user.profile') }}">{{ current_user.first_name or current_user.username or current_user.email }}</a>
+                <a href="{{ url_for('user.profile') }}">{{ current_user.user.first_name or current_user.user.username or current_user.user.email }}</a>
                 &nbsp; | &nbsp;
                 <a href="{{ url_for('user.logout') }}">Sign out</a>
             {% else %}
-                <a href="{{ url_for('user.login') }}">Sign in</a>
+                <a href="{{ url_for('user.get_login') }}">Sign in</a>
             {% endif %}
         </div>
     </div>
@@ -54,7 +54,7 @@
     <h1>DEBUG SITEMAP</h1>
     <h3>Epic</h3>
     <ul>
-        <li><a href="{{ url_for('epic.create') }}">Epic Create</a></li>
+        <li><a href="{{ url_for('epic.get_create') }}">Epic Create</a></li>
         <li><a href="{{ url_for('epic.list_all') }}">Epic List</a></li>
         <li>Epic Show requires Epic ID, see Epic List</li>
     </ul>
@@ -80,7 +80,7 @@
     </ul>
     <h3>Tag</h3>
     <ul>
-        <li><a href="{{ url_for('tag.create') }}">Tag Create</a></li>
+        <li><a href="{{ url_for('tag.get_create') }}">Tag Create</a></li>
         <li><a href="{{ url_for('tag.list_all') }}">Tag List</a></li>
         <li>Tag Show requires Tag ID, See Tag List</li>
     </ul>
