+++ noscrum/templates/epic/list.html
@@ -5,7 +5,7 @@
 {% endblock %}
 
 {% block content %}
-<a class="small button" href="{{ url_for('epic.create') }}">Create New Epic</a>
+<a class="small button" href="{{ url_for('epic.get_create') }}">Create New Epic</a>
 <br>
     <div class="grid-x grid-padding-x">
         {% for epic in epics %}
