+++ noscrum/templates/tag/list.html
@@ -6,7 +6,7 @@
 
 {% block content %}
 {% for tag in tags %}
-<div class="tag" edit_url="{{ url_for('tag.show',tag_id=tag.id) }}">
+<div class="tag" edit_url="{{ url_for('tag.update',tag_id=tag.id) }}">
     {{ tag.tag }}
 </div>
 {% endfor %}
