+++ noscrum/templates/work/read_del.html
@@ -10,7 +10,7 @@
 Hours Worked: {{ work.hours_worked }}<br>
 Work Date: {{ work.work_date }}<br>
 Status: {{ work.status }}<br>
-<form method="delete" actuon="{{ url_for('work.read_delete',work_id=work.id) }}">
+<form method="delete" actuon="{{ url_for('work.delete',work_id=work.id) }}">
     <button class="button" type="submit" value="Delete" value="delete-work">Delete</button>
 </form>
 {% endblock %}
