+++ requirements.txt
@@ -7,7 +7,7 @@ click==7.1.2
 cryptography==36.0.1
 dnspython==2.2.0
 email-validator==1.1.3
-Flask==1.1.2
+Flask
 Flask-Babel==2.0.0
 Flask-BabelEx==0.9.4
 Flask-Foundation==2.1
